+++ internal/table/retry.go
@@ -62,7 +62,7 @@ func doTx(ctx context.Context, c SessionProvider, op table.TxOperation, opts tab
 			err = op(ctx, tx)
 
 			if err != nil {
-				return err
+				return errors.WithStackTrace(err)
 			}
 
 			_, err = tx.CommitTx(ctx, opts.TxCommitOptions...)
@@ -96,7 +96,13 @@ func do(ctx context.Context, c SessionProvider, op table.Operation, opts table.O
 				attempts++
 			}()
 
-			return op(ctx, s)
+			err = op(ctx, s)
+
+			if err != nil {
+				return errors.WithStackTrace(err)
+			}
+
+			return nil
 		},
 	)
 }
@@ -133,9 +139,9 @@ func SingleSession(s Session) SessionProvider {
 }
 
 var (
-	errNoSession         = fmt.Errorf("no session")
-	errUnexpectedSession = fmt.Errorf("unexpected session")
-	errSessionOverflow   = fmt.Errorf("session overflow")
+	errNoSession         = errors.New(fmt.Errorf("no session"))
+	errUnexpectedSession = errors.New(fmt.Errorf("unexpected session"))
+	errSessionOverflow   = errors.New(fmt.Errorf("session overflow"))
 )
 
 type singleSession struct {
@@ -236,11 +242,11 @@ func retryBackoff(
 			}
 
 			if !m.MustRetry(isOperationIdempotent) {
-				return
+				return errors.WithStackTrace(err)
 			}
 
 			if retry.Wait(ctx, fastBackoff, slowBackoff, m, i) != nil {
-				return
+				return errors.WithStackTrace(err)
 			}
 
 			code = m.StatusCode()