+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+* Refactored `cluster.Cluster` and `balancer.Balancer` interfaces (removed `Update` method)
+* Replaced `cluster.Update` with `cluster.Remove` and `cluster.Insert` calls
+* Removed `trace.Driver.OnClusterUpdate` event
+* Fixed bug with unexpected changing of local datacenter flag in endpoint
+* Refactored errors wrapping (stackedError are not ydb error now, checking `errors.IsYdb(err)` with `errors.As` now)
+* Wrapped retry operation errors with `errors.WithStackTrace(err)`
+* Changed `trace.RetryLoopStartInfo.Context` type from `context.Context` to `*context.Context`
+
 ## v3.16.9
 * Refactored internal operation and transport errors